+++ API/DTOs/NewPaymentDto.cs
@@ -4,7 +4,7 @@ namespace API.DTOs
 {
     public class NewPaymentDto
     {
-        [Required] public int Amount { get; set; }
+        [Required] public double Amount { get; set; }
         [Required] public int PayMonth { get; set; }
         [Required] public string Method { get; set; }
         [Required] public DateOnly PayDate { get; set; }
